requirements.txt
amqp
PyCryptodome
pycocotools==2.0.0
mysqlclient
torch==1.8.0
torchvision==0.9.0
onnx==1.10.1
Pillow==8.2.0
coremltools==5.1.0
django==3.2.3
celery==5.0.5
rabbitmq
numpy==1.21.4
django-sslserver
boto3
azure-storage-blob
